LayoutTests/fast/canvas/canvas-too-large-to-draw-expected.txt

CONSOLE MESSAGE: line 35: Canvas area exceeds the maximum limit (width * height > 268435456).
 
